rac dung ai care

+++ src/main/java/com/example/project2/entities/Entity.java
@@ -10,15 +10,15 @@ import javafx.scene.paint.Color;
 
 public abstract class Entity {
     //Tọa độ ngang tính từ góc trái trên trong Canvas
-    protected int x;
+    protected double x;
 
     //Tọa độ dọc tính từ góc trái trên trong Canvas
-    protected int y;
+    protected double y;
 
     protected Image img;
 
     //Khởi tạo đối tượng, chuyển từ tọa độ đơn vị sang tọa độ trong canvas
-    public Entity(int xUnit, int yUnit, Image img) {
+    public Entity(double xUnit, double yUnit, Image img) {
         this.x = xUnit * Sprite.SCALED_SIZE;
         this.y = yUnit * Sprite.SCALED_SIZE;
         this.img = img;
